Move TopAppBar icons into dedicated file
