+++ library-compose/src/main/java/com/spendesk/grapes/compose/appbar/GrapesTopAppBar.kt
@@ -8,23 +8,16 @@ import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.Text
 import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.spendesk.grapes.compose.R
 import com.spendesk.grapes.compose.theme.GrapesTheme
 
 /**
@@ -39,7 +32,7 @@ fun GrapesTopAppBar(
     modifier: Modifier = Modifier,
     navigationIcon: @Composable (() -> Unit)? = null,
     actions: @Composable RowScope.() -> Unit = {},
-    elevationDp: Dp = GrapesTheme.dimensions.elevationNormal
+    elevationDp: Dp = GrapesTheme.dimensions.elevationNormal,
 ) {
     TopAppBar(
         modifier = modifier,
@@ -76,7 +69,7 @@ fun GrapesTopAppBar(
 @Composable
 fun GrapesTopAppBarTitle(
     title: String,
-    modifier: Modifier = Modifier
+    modifier: Modifier = Modifier,
 ) {
     Text(
         modifier = modifier,
@@ -85,74 +78,6 @@ fun GrapesTopAppBarTitle(
         maxLines = 2,
         overflow = TextOverflow.Ellipsis,
         style = GrapesTheme.typography.titleL,
-    )
-}
-
-@Composable
-fun GrapesTopAppBarTitleWithSubtitle(
-    title: String,
-    subTitle: String,
-    modifier: Modifier = Modifier
-) {
-    Column(modifier = modifier) {
-        GrapesTopAppBarTitle(
-            title = title,
-        )
-
-        Text(
-            text = subTitle,
-            lineHeight = GrapesTheme.typography.bodyM.fontSize,
-            style = GrapesTheme.typography.bodyM,
-        )
-    }
-}
-
-@Composable
-fun GrapesTopAppBarIconButton(
-    icon: @Composable () -> Unit,
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit = {}
-) {
-    IconButton(
-        modifier = modifier,
-        content = icon,
-        onClick = onClick,
-    )
-}
-
-@Composable
-fun GrapesTopAppBarBackIcon(
-    modifier: Modifier = Modifier,
-) {
-    Icon(
-        modifier = modifier.size(24.dp),
-        imageVector = ImageVector.vectorResource(R.drawable.ic_arrow_back),
-        contentDescription = stringResource(id = R.string.grapes_top_app_bar_back_icon_description),
-        tint = GrapesTheme.colors.structureComplementary,
-    )
-}
-
-@Composable
-fun GrapesTopAppBarCloseIcon(
-    modifier: Modifier = Modifier,
-) {
-    Icon(
-        modifier = modifier.size(24.dp),
-        imageVector = ImageVector.vectorResource(R.drawable.ic_close),
-        contentDescription = stringResource(id = R.string.grapes_top_app_bar_close_icon_description),
-        tint = GrapesTheme.colors.structureComplementary
-    )
-}
-
-@Composable
-fun GrapesTopAppBarMoreIcon(
-    modifier: Modifier = Modifier,
-) {
-    Icon(
-        modifier = modifier.size(24.dp),
-        imageVector = ImageVector.vectorResource(R.drawable.ic_more_vertical),
-        contentDescription = stringResource(id = R.string.grapes_top_app_bar_more_icon_description),
-        tint = GrapesTheme.colors.structureComplementary
     )
 }
 
@@ -176,16 +101,6 @@ fun GrapesSimpleTopAppBarPreview() {
             GrapesTopAppBar(
                 modifier = Modifier.fillMaxWidth(),
                 title = { GrapesTopAppBarTitle(title = "Top app bar  with a very very very very very very very long title") },
-            )
-
-            GrapesTopAppBar(
-                modifier = Modifier.fillMaxWidth(),
-                title = {
-                    GrapesTopAppBarTitleWithSubtitle(
-                        title = "Top app bar title",
-                        subTitle = "Top app bar subtitle"
-                    )
-                },
             )
 
             GrapesTopAppBar(
@@ -248,45 +163,7 @@ fun GrapesSimpleTopAppBarPreview() {
 
             GrapesTopAppBar(
                 modifier = Modifier.fillMaxWidth(),
-                title = {
-                    GrapesTopAppBarTitleWithSubtitle(
-                        title = "Top app bar title",
-                        subTitle = "Top app bar subtitle"
-                    )
-                },
-                navigationIcon = {
-                    GrapesTopAppBarIconButton(
-                        icon = { GrapesTopAppBarBackIcon() },
-                        onClick = { Log.i("GrapesTopAppBar", "navigationIcon click") }
-                    )
-                },
-            )
-
-            GrapesTopAppBar(
-                modifier = Modifier.fillMaxWidth(),
                 title = { GrapesTopAppBarTitle(title = "Top app bar title") },
-                navigationIcon = {
-                    GrapesTopAppBarIconButton(
-                        icon = { GrapesTopAppBarBackIcon() },
-                        onClick = { Log.i("GrapesTopAppBar", "navigationIcon click") }
-                    )
-                },
-                actions = {
-                    GrapesTopAppBarIconButton(
-                        icon = { GrapesTopAppBarMoreIcon() },
-                        onClick = { Log.i("GrapesTopAppBar", "action button click") }
-                    )
-                },
-            )
-
-            GrapesTopAppBar(
-                modifier = Modifier.fillMaxWidth(),
-                title = {
-                    GrapesTopAppBarTitleWithSubtitle(
-                        title = "Top app bar title",
-                        subTitle = "Top app bar subtitle"
-                    )
-                },
                 navigationIcon = {
                     GrapesTopAppBarIconButton(
                         icon = { GrapesTopAppBarBackIcon() },
